compiler error fix

+++ stp/protocol/timers.go
@@ -142,7 +142,7 @@ func (p *StpPort) DecrementTimerCounters() {
 	// prt owner
 	if p.RrWhileTimer.count > 0 {
 		if p.PrtMachineFsm.Machine.Curr.CurrentState() == PrtStateRootPort {
-			if p.RrWhileTimer.count != p.b.RootTimes.ForwardingDelay &&
+			if p.RrWhileTimer.count != int32(p.b.RootTimes.ForwardingDelay) &&
 				p.Selected &&
 				!p.UpdtInfo {
 				p.PrtMachineFsm.PrtEvents <- MachineEvent{
